adk-agents/requirements.txt
google-generativeai>=0.3.0
supabase>=2.0.0
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
functions-framework>=3.0.0 